shared_lib/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shared"
version = "1.0.0"
description = "Shared library for Python Discord bots"
authors = [
    {name = "Sawyer", email = "sawyer@example.com"},
]
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.4.0",
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "structlog>=24.4.0",
    "prometheus-client>=0.20.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/shared"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "B", "I", "N", "UP", "ANN", "S", "C4", "T20"]
ignore = ["ANN101", "ANN102", "S101"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "discord.*"
ignore_missing_imports = true